Estimate the MBE posterior distribution CE with only the two surrogate endpoints

Uses the estimated treatment effects on the two surrogate endpoints of a new trial and the historical posterior draws to estimate the posterior distribution of MBE.

Usage

surrogate_MBE(
  mcmc_dat,
  sample_dat,
  diffuse = TRUE,
  diffuse_se = 100,
  intercept0 = TRUE
)

Arguments

mcmc_dat

A data frame or matrix containing posterior draws from the two-surrogate historical model. Must contains columns of the variables named: alphaCEonSur1Sur2, b1CEonSur1Sur2, b2CEonSur1Sur2, and SigSqCEonSur1Sur2.

sample_dat

A named list containing exactly nine scalar values: Sur1Est, Sur1SE, Sur2Est, Sur2SE, and R12.

diffuse

A logical value indicating whether to use a diffuse prior. Defaults to TRUE.

diffuse_se

A positive numeric value specifying the standard deviation of the diffuse prior for the two surrogate effects. Only used when diffuse = TRUE. defaults to 100.

intercept0

A logical value indicating whether to center the intercept term of the meta-regression to zero. Defaults to TRUE.

Value

A data table containing:

post_Sigma22: Posterior draws of the variance of the first surrogate.

post_Sigma23: Posterior draws of the covariance between the first and second surrogates.

post_Sigma33: Posterior draws of the variance of the second surrogate.

post_mu2: Posterior draws of the mean of the first surrogate.

post_mu3: Posterior draws of the mean of the second surrogate.

post_CE: Posterior draws of the treatment effect on the clinical endpoint.

post_psi2: Posterior draws of the treatment effect on the first surrogate.

post_psi3: Posterior draws of the treatment effect on the second surrogate.

Details

Let \(\hat{\gamma}_0 = (\hat{\gamma}_{0, 1}, \hat{\gamma}_{0, 2})^T\) be the vector of treatment effects on the two surrogate endpoints of a new trial, and let \(\hat{\gamma}_0 \sim N(\gamma_0, \Sigma_0)\) where \(\Sigma_0\) is assumed to be known (estimated from the new trial data). We model prior on \(\gamma_0 = (\gamma_{0, 1}, \gamma_{0, 2})^T\), the vector of true treatment effects on the two surrogate endpoints. The posterior distribution of \(\gamma_0\) is normal by conjugacy. The prior distribution can either be estimated from the historical posterior draws or be diffuse. For diffuse, a bivariate normal distribution with mean vector 0 and covariance matrix diffuse_se^2 \( \times I_2\) is used. The bigger the diffuse_se, the more diffuse the prior is. Samples of the posterior distribution of \(\gamma_0\) are obtained and then used to estimate the posterior distribution of MBE on the clinical endpoint using the relationship: $$\theta_0 = \alpha_\theta + \beta_{\gamma_1} \cdot \gamma_{0, 1} + \beta_{\gamma_2} \cdot \gamma_{0, 2} + \epsilon, \quad \epsilon \sim N(0, \lambda^2_\theta).$$

Historical posterior distribution of \(\alpha_\theta, \beta_{\gamma_1}, \beta_{\gamma_2}, \lambda^2_\theta\) is used to estimate the posterior distribution of MBE on the clinical endpoint without updating its distribution.

Examples

# Use the example data provided in the package.
# Load the historical posterior samples where the model was fit
# with random intercept and inverse gamma prior for variance parameters
# on the 65/66 simulated trials in `trial_sim_dat` dataset.
# It's the first trial that was left out as if it was a new trial.
# We demonstrate estimating the posterior distribution of MBE on this first
# row that was left out.
data("historical_posterior")
data("trial_sim_dat")
one_sim_dat<-list(
# treatment effect on chronic slope
Sur1Est=trial_sim_dat$Sur1_est[1],
Sur1SE=trial_sim_dat$Sur1_se[1],

# treatment effect on acute slope
Sur2Est=trial_sim_dat$Sur2_est[1],
Sur2SE=trial_sim_dat$Sur2_se[1],

# correlation between chronic slope and acute slope
R12=trial_sim_dat$Cor_Sur1_Sur2[1]
)

MBE_from_surrogate_only<-surrogate_MBE(
mcmc_dat = historical_posterior,
sample_dat=one_sim_dat,
diffuse = TRUE, #We will not use the historical posterior
#distribution of the two surrogate endpoints to estimate the
#posterior distribution of MBE. Instead, we will use a diffuse prior.
diffuse_se = 100,
intercept0 = FALSE)